Setup Snowpack
Default configuration is 'development', using the 'build' script you
may pass a '--production' or '-p' for short option to enable the
'production' build mode. The 'build' script builds the app in the
'dist' folder. In 'production' mode the content will be minified and
bundled. The 'dev' script starts a dev-server at the 8080 port. All the
scripts are stored in the 'scripts' folder.

All folders are mounted as-is except 'www' which is mounted at the
root. There is a couple of those folders: 'components',
'styles', 'app', 'assets', 'themes', 'utils', and previously
described 'www'.

The 'components' folder contains '.tsx' files
representing a component having the same name and exporting it as
a default export. The 'styles' folder contains '.sass' files attached
to matching '.tsx' file. React is importing by the first line of '.tsx'
file, and the matching style is by the second one. There are also a
couple of subfolders: 'pages', and 'layouts'. Layouts are always the
result of page rendering and they know that they will be stored in the
root element, so, they can use the 'React.Fragment' component as their
rendering result. The 'pages' subfolder also contains an 'index.ts'
file which is imported in the 'App.tsx' file to approach the 'pages'
subfolder like a module. The 'App' component will always be, renders in
the 'www/index.tsx' file, creates initiates Redux, and routes pages.
The 'App.sass' file contains basic CSS preambule and a couple of
CSS classes for better DX: 'hidden' and 'collapsed' which are
representing the state of a potential user. There may be some conflicts
between local class and global class, so, class users should use
SASS anchors to explicitly override the behaviour.

A 'store.ts' file, the core of Redux, is stored in the 'app' folder and
have a default export, the Redux store. The 'slices' subfolder contains
slices named '<file name>Slice' and exported as a default export. Same
with the 'sagas' and 'hooks' folder. There is also an 'api' Each of
these folders also contains one 'index.ts' file used by the 'store.ts'
file.