Setup slices
'photos', 'headerPhoto', 'searchFilter', 'user', and 'errors' slices
have been configured.

The main content of the 'photos' slice is a photo dictionary, and their
order (ID array). It also contains a page capacity that can be
configured only when there is no downloading completed and represents
amount of downloading photos per page.

The 'headerPhoto' slice is used by 'DefaultLayout' to store its header
photo. It has the same downloading mechanics as the 'photos' slice.

The 'searchFilter' slice contains filters for the 'SearchPage' such as
'query', 'orientation', 'size', and 'color'. Changes should be
represented in the URL's query of the 'SearchPage'. Icons of
'orientation' and 'size' filters are stored in the matching assetss
subfolders.

The 'user' slice contains stored in local storage user-specific data
such as 'language' and 'likes'. The 'language' property is a language
used in a 'assets/localization.json' file.

The 'errors' slice is just a list of errors which should be shown to
the user as notifications. As these errors are unique objects, they
are identified by their own references.