Setup hooks
The biggest part of all the hooks are 'use<slice name>' hooks which are
depended on the 'useSlice' hook. It takes a slice name and a selector
for its state. The first returned value is selected state, and the
second one is a used slice. Used slices are object containing slice
state, slice name, and also a 'dispatch' method which accepts type of
action as a first arg and its payload as a second arg.

There is also a 'useScrollPosition' hook returning scroll position
snapshots. Scroll position snapshot is an object containing scorll
position in pixels and percents, 2 flags indicating is scroll in one
of the extreme positions, and a helper function helps to determine has
the scroll reached some position or not.