Add color transformation utils
`Color` type is an array containing 3 color channels specified by
type user. `rgbToHsv` and `hsvToRgb` accepts and returns `Color`
instances and transforms it in different color models. There is also
a `parseRgbColor` function, parses hex colors ('#rrggbb' or '#rgb')
to a `Color` instance. And `formatRgbColor` does the opposite, but
always formats as '#rrggbb'.